api/pyproject.toml
[tool.poetry]
name = "rogaining-system-back"
version = "0.1.0"
description = "rogaining_systemのlambdaデプロイ用のプロジェクト"
authors = ["sion908 <sionn908@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.101.0"
mangum = "^0.17.0"
sqlalchemy = "^2.0.30"
pymysql = "^1.1.0"
aiomysql = "^0.2.0"
cryptography = "^41.0.3"
sqlalchemy-utils = "^0.41.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
uvicorn = "^0.23.2"
line-bot-sdk = "^3.7.0"
boto3 = "^1.34.36"
jinja2 = "^3.1.3"
fastapi-csrf-protect = "^0.3.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic-settings = "^2.2.1"


[tool.poetry.group.dev.dependencies]
alembic = "^1.13.1"
python-dateutil = "^2.8.2"
requests = "^2.31.0"
httpx = "^0.25"
aiosqlite = "^0.19.0"
uvicorn = "^0.23.2"
setuptools = "^68.1.2"
pyyaml = "^6.0.1"
asyncpg = "^0.29.0"
httpx-auth = "^0.18.0"
fastapi-to-openapi = "0.2.2"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "0.23.3"
pytest-mock = "^3.12.0"
freezegun = "^1.4.0"


[tool.poetry.group.linter.dependencies]
mypy = "^1.7.1"
black = "^23.11.0"
isort = "^5.12.0"
flake8-isort = "^6.1.1"
flake8-bugbear = "^23.12.2"
flake8-builtins = "^2.2.0"
flake8-eradicate = "^1.5.0"
flake8-unused-arguments = "^0.0.13"
flake8-pytest-style = "^1.7.2"
pyproject-flake8 = "^6.1.0"


[tool.poetry.group.alembic.dependencies]
alembic = "^1.13.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
line_length = 120
multi_line_output = 3
include_trailing_comma = true
known_local_folder=['config',]
skip = ["src/database/migrations"]

[tool.black]
line-length = 80

[tool.flake8]
max-line-length = 100
max-complexity = 10
ignore = "E203,E402,E266,W503,B008,A003"
exclude = "__pycache__,tests,.venv"

[tool.mypy]
python_version = "3.9"
no_strict_optional = true
ignore_missing_imports = true
check_untyped_defs = true

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = ["tests",]
filterwarnings = ["ignore::DeprecationWarning",]
asyncio_mode = "auto"
